+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -250,7 +250,14 @@ public class Settings extends Config {
 	public static boolean preventTradingWithOwnShop = true;
 	public static boolean preventTradingWhileOwnerIsOnline = false;
 	public static boolean useStrictItemComparison = false;
+
 	public static boolean incrementVillagerStatistics = false;
+	public static boolean simulateVillagerTradingSounds = true;
+	public static boolean simulateVillagerAmbientSounds = false;
+	public static boolean simulateWanderingTraderTradingSounds = true;
+	public static boolean simulateWanderingTraderAmbientSounds = false;
+	public static boolean simulateTradingSoundsOnlyForTheTradingPlayer = true;
+
 	public static SoundEffect tradeSucceededSound = new SoundEffect(Sound.UI_BUTTON_CLICK).withPitch(2.0f).withVolume(0.3f);
 	public static SoundEffect tradeFailedSound = new SoundEffect(Sound.BLOCK_BARREL_CLOSE).withPitch(2.0f).withVolume(0.5f);